Add gcloud plugin (#8144)
* A gcloud plugin

This PR addresses issue #6205

This adds support for loading completion for the Google Cloud SDK
command line tools. It searches the known paths for an SDK and loads the
provided completion if it is found. Users can supply a custom location
for the SDK by setting `CLOUDSDK_HOME` in their `zshrc` before loading
oh-my-zsh plugins.

* Canoncial zsh and some safe guards

Based on a PR review from mcornella. All off this has now been tested
on the following variants:

  - macOS 10.14.6 - Homebrew
  - macOS 10.14.6 - Custom install
  - Ubuntu 16.04.6 LTS - apt install
  - Ubuntu 18.04.3 LTS - apt install
  - Ubuntu 18.04.3 LTS - snap classic install
  - CentOS 7 - yum install